Chỉnh sửa cho phù hợp vs điều chỉnh trong database

+++ src/main/java/com/swp391_8/schoolhealth/controller/StudentController.java
@@ -21,18 +21,16 @@ public class StudentController {
     public ResponseEntity<List<Student>> getAllStudents() {
         List<Student> students = studentService.getAllStudents();
         return ResponseEntity.ok(students);
-    }
-
-    @GetMapping("/{id}")
+    }    @GetMapping("/{id}")
     @PreAuthorize("hasRole('MEDICAL_STAFF') or hasRole('ADMIN') or @securityService.isParentOfStudent(authentication, #id)")
-    public ResponseEntity<Student> getStudentById(@PathVariable Long id) {
+    public ResponseEntity<Student> getStudentById(@PathVariable Integer id) {
         Student student = studentService.getStudentById(id);
         return ResponseEntity.ok(student);
     }
 
     @GetMapping("/parent/{parentId}")
     @PreAuthorize("hasRole('MEDICAL_STAFF') or hasRole('ADMIN') or authentication.principal.id == #parentId")
-    public ResponseEntity<List<Student>> getStudentsByParentId(@PathVariable Long parentId) {
+    public ResponseEntity<List<Student>> getStudentsByParentId(@PathVariable Integer parentId) {
         List<Student> students = studentService.getStudentsByParentId(parentId);
         return ResponseEntity.ok(students);
     }
@@ -42,18 +40,16 @@ public class StudentController {
     public ResponseEntity<Student> createStudent(@RequestBody Student student) {
         Student newStudent = studentService.createStudent(student);
         return ResponseEntity.ok(newStudent);
-    }
-
-    @PutMapping("/{id}")
+    }    @PutMapping("/{id}")
     @PreAuthorize("hasRole('MEDICAL_STAFF') or hasRole('ADMIN')")
-    public ResponseEntity<Student> updateStudent(@PathVariable Long id, @RequestBody Student student) {
+    public ResponseEntity<Student> updateStudent(@PathVariable Integer id, @RequestBody Student student) {
         Student updatedStudent = studentService.updateStudent(id, student);
         return ResponseEntity.ok(updatedStudent);
     }
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<?> deleteStudent(@PathVariable Long id) {
+    public ResponseEntity<?> deleteStudent(@PathVariable Integer id) {
         studentService.deleteStudent(id);
         return ResponseEntity.ok().build();
     }
