+++ src/main/java/com/swp391_8/schoolhealth/model/CheckupResult.java
@@ -12,10 +12,9 @@ import java.time.LocalDate;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CheckupResult {
-    @Id
+public class CheckupResult {    @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private LocalDate checkupDate;
