+++ src/main/java/com/swp391_8/schoolhealth/repository/BlogPostRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BlogPostRepository extends JpaRepository<BlogPost, Long> {
-    List<BlogPost> findByAuthorId(Long authorId);
+public interface BlogPostRepository extends JpaRepository<BlogPost, Integer> {
+    List<BlogPost> findByAuthorId(Integer authorId);
     List<BlogPost> findAllByOrderByCreatedAtDesc();
 }
